Add SimpleListener to JellyfishNotification
This adds SimpleListener to JellyNotification module. Included also is a
basic spec that mocks the Jellyfish Project model and attempts to send
an email notification in test and validate that the from email matches
the from user specified in the application dotenv on the event a
project_creation_successful is broadcasted thru Wisper.