Update Simple Listener Spec
Pass a Project OpenStruct Mock object to the listener to use for
processing. Project mock as a JSON string which will be parsed into a
hash to simulate response from Project controller in JF api.
Changed broadcast to publish to reflect current syntax and passed it an
empty string to invoke Action Mailer configuration settings initialized
for the test to pull up a dummy to and from address. Also abstracted
out what happens when mail is delivered asynchronously. Need to figure
out how to better test asynchronous mail with delayed jobs.