Updated spec for approval update
The object fed to the Wisper publisher differs based on the controller.
We can pass the project object from the Project Approvals Controller on
update, but the project object is not available with an id from the
Project Controller on create so need to pass a stubbed response object
instead.