Add publish order create notifications to listener
Added a response to published order create notifications that will send
and email containing each of the order items in the order to the passed
recipients (or to default recipient specified in dotenv if var empty).
Spec that was added contains a rather large mocked order object which
may need to be simplified going forward.